:sparkles: feat : DTO 변수타입 변경과 etf_nav 과거 데이터 snapshot에 미리 저장 BatchConfig 종료후로 변경

+++ src/main/java/com/finsight/backend/domain/vo/user/HoldingsVO.java
@@ -4,7 +4,6 @@ import com.finsight.backend.dto.request.TradeRequest;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -14,7 +13,7 @@ public class HoldingsVO {
     private String userId;
     private String productCode;
     private String productCategory;
-    private BigDecimal holdingsTotalPrice;
+    private Double holdingsTotalPrice;
     private Integer holdingsTotalQuantity;
     private String holdingsStatus; // holding, zero, terminated, expired
